pyproject.toml
[project]
name = "findmy-setup"
version = "0.1.0"
description = ""
authors = [
    {name = "HermanL02",email = "Hermanyiqunliang@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
    "findmy (>=0.9.4,<0.10.0)",
    "pyicloud (>=2.1.0,<3.0.0)",
    "pymongo (>=4.15.3,<5.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
